// src/types/zone.ts
export interface Zone {
  slug: string
  name: string
  lat: number
  lng: number
  scoreLCD: number
  scoreLMNP: number
  tension: number
  transport: number
  commerce: number
  population: number
  prixM2: number
  loyerM2: number
  nuiteeMin: number
  nuiteeMax: number
  occupMoy: number
  description: string
  why: string
  regulation: string
  tips: string
}